fix loi

+++ Doanh Nghiep/SubForm/Dang_ky_moi/frm_dang_ky_moi.cs
@@ -347,6 +347,23 @@ namespace QuanLyDangKyDoanhNghiep.Doanh_Nghiep.Dang_ky_moi
                 MessageBox.Show("Hồ sơ đã nộp không được phép hủy");
                 return;
             }
+            using (QuanLyDangKyDoanhNghiepEntities db = new QuanLyDangKyDoanhNghiepEntities())
+            {
+                doanh_Nghiep = db.doanh_nghiep.Where(item => item.id == doanh_Nghiep.id).FirstOrDefault();
+                doanh_Nghiep.loai_hinh = null;
+                doanh_Nghiep.id_dia_chi = null;
+                doanh_Nghiep.thongtin_diachi = null;
+                doanh_Nghiep.id_chu_so_huu = null;
+                doanh_Nghiep.nhan_su_doanh_nghiep = null;
+                doanh_Nghiep.id_nguoi_dai_dien_phap_luat = null;
+                doanh_Nghiep.nhan_su_doanh_nghiep1 = null;
+                doanh_Nghiep.nhan_su_doanh_nghiep2 = null;
+                doanh_Nghiep.thong_tin_thue = null;
+                doanh_Nghiep.id_thong_tin_thue = null;
+                doanh_Nghiep.nganh_nghe_dangky = null;
+                db.Entry(doanh_Nghiep).State = System.Data.Entity.EntityState.Modified;
+                db.SaveChanges();
+            }
             xoa_nganh_nghe_dang_ky();
             xoa_nhan_su_doanh_nghiep();
             xoa_dia_chi_doanh_nghiep();
@@ -362,9 +379,10 @@ namespace QuanLyDangKyDoanhNghiep.Doanh_Nghiep.Dang_ky_moi
                 using (QuanLyDangKyDoanhNghiepEntities db = new QuanLyDangKyDoanhNghiepEntities())
                 {
                     thongtin_diachi thongtin_Diachi = db.thongtin_diachi.Where(item => item.id == doanh_Nghiep.id_dia_chi).FirstOrDefault();
-                    
-                    db.thongtin_diachi.Remove(thongtin_Diachi);
-                    db.SaveChanges();
+                    if (thongtin_Diachi != null) { 
+                        db.thongtin_diachi.Remove(thongtin_Diachi);
+                        db.SaveChanges();
+                    }
                 }
             }
         }
@@ -401,7 +419,7 @@ namespace QuanLyDangKyDoanhNghiep.Doanh_Nghiep.Dang_ky_moi
                 List<nhan_su_doanh_nghiep> nhan_Su_Doanh_Nghieps = db.nhan_su_doanh_nghiep.Where(item => item.id_doanh_nghiep == doanh_Nghiep.id).ToList<nhan_su_doanh_nghiep>();
                 foreach (nhan_su_doanh_nghiep item in nhan_Su_Doanh_Nghieps)
                 {
-                    
+                        
                         db.nhan_su_doanh_nghiep.Remove(item);
                         db.SaveChanges();
                 }
@@ -412,6 +430,7 @@ namespace QuanLyDangKyDoanhNghiep.Doanh_Nghiep.Dang_ky_moi
         {
             using (QuanLyDangKyDoanhNghiepEntities db = new QuanLyDangKyDoanhNghiepEntities())
             {
+                doanh_Nghiep = db.doanh_nghiep.Where(item => item.id == doanh_Nghiep.id).FirstOrDefault();
                 db.doanh_nghiep.Remove(doanh_Nghiep);
                 db.SaveChanges();
             }
